Fix/improve server swap (#315)
* remove unnecessary TODO

* get rid of wallet_program_router_account

* parse deadline

* add fee token specification in Opportunity

* add referral fees to quot
e api & opportunity

* change location of TODO

* rename bid payment type

* nit improve todo

* feat: create quote response submit method, typescript (#312)

* create quote response submit method, typescript

* ignore generated files in linting

* bump package version

* convert to date

* add todo

* make referral_fee_bps optional

* fix the optional/default behavior in server + client

* address comments 1

* add a todo around distinguishing enum variants

+++ contracts/svm/programs/express_relay/src/lib.rs
@@ -380,7 +380,7 @@ pub struct SwapArgs {
     pub amount_input:     u64,
     pub amount_output:    u64,
     // The referral fee is specified in basis points
-    pub referral_fee_bps: u64,
+    pub referral_fee_bps: u16,
     // Token in which the fees will be paid
     pub fee_token:        FeeToken,
 }
